Refactor PenguinMap component for better maintainability
- Split large PenguinMap.tsx into smaller components
- Created separate files for styles, utilities, and hooks
- Improved code organization and readability
- Separated concerns: UI, data fetching, and rendering logic
